Implement controller interface, fix indirect addressing edge case.
Implemented controller as an interface. This allows flexibility in how the
emulator reads controller input (e.g., keyboard, touch screen, usb controller,
etc.). Fixed indirect addressing bug in which, if the word is on the last
byte of a page (0x..ff), then it must wrap around to the beginning of the page.